bigger task/boards sizes for tablets
So helpfull for me! 
We put a tablet with chrome in fullscreen in the middle of the workshop. 
When a worker start/end a job, they move the task to the corresponding board with a simple move in touchscreen.
With this workflow, we get better control about time we spent in each task.